backend/requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
Pillow>=10.1.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
boto3>=1.34.0
django-storages>=1.14.0
gunicorn>=21.2.0
dj-database-url>=2.1.0
python-telegram-bot>=20.0
